requirements.txt
h5py==2.10.0
imageio==2.8.0
imageio-ffmpeg==0.4.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib==3.1.0
numpy==1.18.5
opencv-contrib-python==4.4.0.44
opencv-contrib-python-headless==4.2.0.32
opencv-python==4.0.0.21
pandas==0.24.2
Pillow==7.0.0
scikit-image==0.14.2
scikit-learn==0.21.3
scipy==1.2.1
seaborn==0.10.0
slim==0.1
tensorboard==1.12.2
tensorflow-estimator==1.15.1
tensorflow-gpu==1.14.0
tqdm==4.46.0
